backend/requirements.txt
fastapi[all]>=0.104.0
uvicorn[standard]>=0.24.0
sqlalchemy>=2.0.23
alembic>=1.12.0
pymupdf>=1.23.0
requests>=2.31.0
beautifulsoup4>=4.12.0
lxml>=4.9.0
feedparser>=6.0.10
python-multipart>=0.0.6
python-dotenv>=1.0.0
aiofiles>=23.2.1
httpx>=0.25.0
google-genai>=1.0.0
anthropic>=0.8.0
openai>=1.3.0
pydantic>=2.4.0
pydantic-settings>=2.0.0
python-dateutil>=2.8.2
